requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langchain==0.1.16
langchain-google-genai==0.0.5
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6